+++ Examples/SwiftUIApp/App.swift
@@ -1,15 +1,27 @@
-import SwiftUI
 import Aptabase
+import SwiftUI
+
+@main
+struct HelloWorldApp: App {
+    var body: some Scene {
+        WindowGroup {
+            CounterView()
+        }
+    }
+}
+
+import Aptabase
+import SwiftUI
 
 struct CounterView: View {
     @State var count: Int = 0
-    
+
     var body: some View {
         VStack {
             Text("Count = \(count)")
             Button(action: {
-                self.count += 1
-                Aptabase.shared.trackEvent("Increment", with: ["count": self.count])
+                count += 1
+                Metrics.shared.track("Increment", ["count": count])
             }) {
                 Text("Increment")
             }.padding()