+++ Package.swift
@@ -4,22 +4,22 @@
 import PackageDescription
 
 let package = Package(
-    name: "Aptabase",
+    name: "AptabaseNomad",
     platforms: [
-      .iOS(.v13),
-      .macCatalyst(.v13),
-      .macOS(.v10_15),
-      .watchOS(.v6),
-      .tvOS(.v13),
+        .iOS(.v16),
+        .macOS(.v13),
+        .watchOS(.v6),
+        .tvOS(.v13),
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "Aptabase",
-            targets: ["Aptabase"]),
+            name: "AptabaseNomad",
+            targets: ["Aptabase"],
+        ),
     ],
     dependencies: [
-      .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.3.0")
+        .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.3.0"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -27,11 +27,12 @@ let package = Package(
         .target(
             name: "Aptabase",
             resources: [
-                .copy("PrivacyInfo.xcprivacy")
-            ]),
+                .copy("PrivacyInfo.xcprivacy"),
+            ],
+        ),
         .testTarget(
             name: "AptabaseTests",
-            dependencies: ["Aptabase"]
-        )
-    ]
+            dependencies: ["Aptabase"],
+        ),
+    ],
 )